% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_plotly_axis.R
\name{adjust_plotly_axis}
\alias{adjust_plotly_axis}
\title{Adjust ticks of second y-axis in plotly}
\usage{
adjust_plotly_axis(y1_data, y2_data)
}
\arguments{
\item{y1_data, y2_data}{Vectors or dataframe columns with the data.}
}
\value{
A \code{list} object with y1 & y2 range min & max and the dticks.
}
\description{
Adjust ticks of second y-axis in plotly
}
\examples{
\dontrun{
ax <- adjust_plotly_axis(mtcars[1], mtcars[2])
plot_ly() \%>\%
plotly_gelayout(hovermode = "closest") \%>\%
  layout(
     yaxis = list(
        overlaying = "y2",
        range = c(ax$y1_range_min, ax$y1_range_max),
        dtick = ax$y1_dtick / n_ticks
        ),
     yaxis2 = list(
        side = "right",
        range = c(ax$y2_range_min, ax$y2_range_max),
        dtick = ax$y2_dtick / n_ticks
        )
     )
}

}
\seealso{
\href{https://github.com/VictorBezak/Plotly_Multi-Axes_Gridlines}{link}
}
\author{
Yannick Schwarz
}
