% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_ge.R
\name{colors_ge}
\alias{colors_ge}
\title{Function to extract GE KVG colors as hex codes}
\usage{
colors_ge(...)
}
\arguments{
\item{...}{Character names of additional colors}
}
\description{
Function to extract GE KVG colors as hex codes
}
\examples{
df <- data.frame(
   year = c(2019, 2020, 2021),
   value = c(4.2, 10, 29.5)
   )

ggplot2::ggplot(data = df, ggplot2::aes(y = value, x = year)) +
   ggplot2::geom_line(color = GEviz::colors_ge("hellrot"))

}
