% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_ge.R
\name{geom_text_ge}
\alias{geom_text_ge}
\title{Text layer for ggplot2}
\usage{
geom_text_ge(label, flip = FALSE)
}
\arguments{
\item{label}{Variable from input object for naming the data points in the plot.}

\item{flip}{Do you flip the coordinates of this plot at the end? If yes, activate flip = TRUE.}
}
\description{
Text layer for ggplot2
}
\examples{
df <- data.frame(
   year = c(2019, 2020, 2021),
   value = c(4.2, 10, 29.5))

ggplot2::ggplot(data = df, ggplot2::aes(x = year, y = value)) +
   ggplot2::geom_bar(stat = "identity") +
   GEviz::geom_text_ge(label = "year")

}
\author{
Yannick Schwarz
}
