% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly_ge.R
\name{ggplotly_ge}
\alias{ggplotly_ge}
\title{GE KVG defaults for ggplotly}
\usage{
ggplotly_ge(
  ggplot_object,
  legend_h_adjust = -0.1,
  hovermode = "x",
  legend_title_space = NULL,
  legend_title = "",
  ...
)
}
\arguments{
\item{ggplot_object}{An object from \code{ggplot2}.}

\item{legend_h_adjust}{Numeric value which can be positive and negative. Default is -0.1.}

\item{hovermode}{Character vector with default "x". Other possible values are "y", "closest", FALSE, "x unified" and "y unified".}

\item{legend_title_space}{Character vector with white space to align legend title.}

\item{legend_title}{Character vector for the title of the legend.}

\item{...}{Further arguments for \code{ggplotly()}. See \link[plotly]{ggplotly} for more.}
}
\description{
GE KVG defaults for ggplotly
}
