% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_ge.R
\name{palette_ge}
\alias{palette_ge}
\title{Return function to interpolate a ge color palette}
\usage{
palette_ge(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ge_palettes. "main" is default, others are: "rotblau", "rot", "blau".}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette().}
}
\description{
Return function to interpolate a ge color palette
}
