% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_gelayout.R
\name{plotly_gelayout}
\alias{plotly_gelayout}
\title{Plotly layout with GE defaults}
\usage{
plotly_gelayout(
  plotly_object,
  plot_title = "",
  xaxis_title = "",
  yaxis_title = "",
  yaxis2_title = "",
  legend_title = "",
  legend_h_adjust = -0.2,
  hovermode = "x",
  legend_title_space = NULL,
  yaxis2 = FALSE,
  yaxis2_args = NULL,
  font_size = 12,
  ...
)
}
\arguments{
\item{plotly_object}{A plotly object.}

\item{plot_title, xaxis_title, yaxis_title, yaxis2_title, legend_title}{Character vectors with the titles.}

\item{legend_h_adjust}{Numeric value which can be positive and negative. Default is -0.1.}

\item{hovermode}{Character vector with default "x". Other possible values are "y", "closest", FALSE, "x unified" and "y unified".}

\item{legend_title_space}{Character vector with white space to align legend title.}

\item{yaxis2}{Logical if a secondary y-axis is desired. Default is FALSE.}

\item{yaxis2_args}{Further arguments which are passed on to \code{layout}.}

\item{font_size}{Integer with the font size.}

\item{...}{Further arguments for \code{ggplotly()}. See \link[plotly]{ggplotly} for more.}
}
\description{
Plotly layout with GE defaults
}
