% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ge.R
\name{theme_ge}
\alias{theme_ge}
\title{Theme for ggplot2}
\usage{
theme_ge(
  panel.grid.major = "grid.y",
  title.x = FALSE,
  title.y = TRUE,
  sec.axis.y = FALSE,
  sec.axis.y.color = "lavender",
  angle.x = 0,
  text.size = 9,
  axis.text.y = TRUE,
  legend.position = "bottom",
  family = "sans",
  ...
)
}
\arguments{
\item{panel.grid.major}{Character vector. "grid.x" for a vertical grid on the x-axis. "grid.y" for a horizontal grid on the y-axis. "none" for no grid at all.}

\item{title.x, title.y}{Title of the axis if needed.}

\item{sec.axis.y}{Logical if y-axis should have a second axis.}

\item{sec.axis.y.color}{Character vector with the color of the second y-axis. Default is lavender".}

\item{angle.x}{Controll the angle of the labels on the x-axis.}

\item{text.size}{Text size of all text elements in the plot.}

\item{axis.text.y}{Logical if y-axis should have text labels.}

\item{legend.position}{Position of the legend. Default is "bottom". "none" removes the legend.}

\item{family}{Character vector with the font which should be used. Default is "sans" which works fine with pdf and html output. For a clean \code{ggplotly} output use "arial".}

\item{...}{Further arguments to pass on to \link[ggplot2]{theme}}
}
\description{
GE KVG branded theme for ggplot2
}
\examples{
\dontrun{
df <- data.frame(
  year = c(2019, 2020, 2021),
  value = c(4.2, 10, 29.5)
)

ggplot2::ggplot(data = df, ggplot2::aes(x = year, y = value)) +
  ggplot2::geom_bar(stat = "identity") +
  GEviz::theme_ge()
}
}
\author{
Magnus Vieten, Yannick Schwarz
}
